+++ lib/src/state_management/state_management.dart
@@ -1,5 +1,6 @@
 export 'async_value.dart';
 export 'debounce_throttle.dart';
+export 'equality.dart';
 export 'future_shard.dart';
 export 'persistent_shard.dart';
 export 'shard.dart';